refactor(wip): move body/request validation schema to each route file
