+++ src/validation/index.ts
@@ -1,3 +1,4 @@
 export * from './openapi-responses';
-export * from './requests';
 export * from './validators';
+export * from './fields';
+export * from './joi';
